move GameType and GameTypes into shared space for use in backend and frontend
